Refactor event
